Added more metrics, specifically from the operating system related to the network, the connections that are open etc.

+++ src/main/java/com/pxs/reaper/action/ReaperActionJmxMetrics.java
@@ -165,7 +165,7 @@ public class ReaperActionJmxMetrics extends ReaperActionMBeanMetrics {
                 throw new RuntimeException(e);
             }
         };
-        return retryWithIncreasingDelay.retry(function, null, maxRetries, finalRetryDelay);
+        return retryWithIncreasingDelay.retry(function, null, (int) maxRetries, finalRetryDelay);
     }
 
     private MBeanServerConnection getMBeanServerConnection() {
